refactor: study results and learning controllers refactor
study results and learning controllers refactor to api endpoints

Closes #825
