+++ src/Sepes.RestApi/ApiEndpoints/Base/EndpointBase.cs
@@ -1,75 +1,15 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using System.Threading;
-using System.Threading.Tasks;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Cors;
+using Microsoft.AspNetCore.Mvc;
 
 namespace Sepes.RestApi.ApiEndpoints.Base
 {
-    public static class EndpointBase
+
+    [ApiController]
+    [Produces("application/json")]
+    [EnableCors("_myAllowSpecificOrigins")]
+    [Authorize]
+    public class EndpointBase : ControllerBase
     {
-        public static class WithRequest<TRequest>
-        {
-            public abstract class WithResult<TResponse> : ControllerBase
-            {
-                public abstract Task<TResponse> Handle(
-                    TRequest request,
-                    CancellationToken cancellationToken = default
-                );
-            }
-
-            public abstract class WithoutResult : ControllerBase
-            {
-                public abstract Task Handle(
-                    TRequest request,
-                    CancellationToken cancellationToken = default
-                );
-            }
-
-            public abstract class WithActionResult<TResponse> : ControllerBase
-            {
-                public abstract Task<ActionResult<TResponse>> Handle(
-                    TRequest request,
-                    CancellationToken cancellationToken = default
-                );
-            }
-
-            public abstract class WithActionResult : ControllerBase
-            {
-                public abstract Task<ActionResult> Handle(
-                    TRequest request,
-                    CancellationToken cancellationToken = default
-                );
-            }
-        }
-
-        public static class WithoutRequest
-        {
-            public abstract class WithResult<TResponse> : ControllerBase
-            {
-                public abstract Task<TResponse> Handle(
-                    CancellationToken cancellationToken = default
-                );
-            }
-
-            public abstract class WithoutResult : ControllerBase
-            {
-                public abstract Task Handle(
-                    CancellationToken cancellationToken = default
-                );
-            }
-
-            public abstract class WithActionResult<TResponse> : ControllerBase
-            {
-                public abstract Task<ActionResult<TResponse>> Handle(
-                    CancellationToken cancellationToken = default
-                );
-            }
-
-            public abstract class WithActionResult : ControllerBase
-            {
-                public abstract Task<ActionResult> Handle(
-                    CancellationToken cancellationToken = default
-                );
-            }
-        }
     }
 }
