+++ src/Sepes.RestApi/ApiEndpoints/ProcessWorkQueue/Work.cs
@@ -1,27 +1,26 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Sepes.Provisioning.Service.Interface;
 using System.Net.Mime;
 using System.Threading.Tasks;
-using Sepes.Provisioning.Service.Interface;
 
-namespace Sepes.RestApi.Controllers
+namespace Sepes.RestApi.ApiEndpoints.ProcessWorkQueue
 {
     [Route("api/provisioningqueue")]
     [ApiController]
     [Authorize()]
-    public class ProcessWorkQueueController : ControllerBase
+    public class Work : ControllerBase
     {
         readonly IResourceProvisioningService _service;
 
-        public ProcessWorkQueueController(IResourceProvisioningService service)
+        public Work(IResourceProvisioningService service)
         {
             _service = service;
         }
 
         [HttpGet("lookforwork")]
         [Consumes(MediaTypeNames.Application.Json)]
-        [Authorize()]
-        public async Task<IActionResult> Work()
+        public async Task<IActionResult> Handle()
         {
             await _service.DequeueAndHandleWork();
             return new OkResult();
